feat: "Lazy load" VCR to reduce plugin overhead
Defer importing `_vcr` in `plugin.py` until the `vcr` fixture is
invoked with a vcr-marked test. Move `merge_kwargs` helper to utils
so it can be imported independently of VCR modules required by
`_vcr.py`. Guard VCR object type hints using `if TYPE_CHECKING`.
