+++ src/pytest_recording/hooks.py
@@ -1,6 +1,9 @@
 from _pytest.config import Config
-from vcr import VCR
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from vcr import VCR
 
 
-def pytest_recording_configure(config: Config, vcr: VCR) -> None:
+def pytest_recording_configure(config: Config, vcr: "VCR") -> None:
     pass  # pragma: no cover
